[read dev] use size_t instead of int for index

+++ readDevInput/src/copy/read_dev_input.cc
@@ -37,7 +37,7 @@ Result ReadDevInput::get(std::string key) {
   return get(key, 0);
 }
 
-Result ReadDevInput::get(std::string key, int index) {
+Result ReadDevInput::get(std::string key, size_t index) {
   return index < read[key].size() ? read[key][index] : Result();
 }
